moved Toc components to separate folder
