+++ packages/code-du-travail-frontend/src/conventions/texte/toc/index.js
@@ -1,0 +1,5 @@
+import TocLink from "./TocLink";
+import TocListItem from "./TocListItem";
+import TocList from "./TocList";
+
+export { TocLink, TocList, TocListItem };
